Sound! Title screen animations!
We now have sound in the game! Have main menu, level 1, and boss theme, as well as jump, double jump, attack, and hurt SFX
I also worked on animating the title screen! Waterfall, below water fall water effect, and king is now a sprite that has idle so that the main title can get more colors in the palette.